Seperated code into *Ops.hpp and *Tests.hpp Renamed MatrixTest.hpp to MatrixTests.hpp
